Rename Airbrakes folder to airbrakes
