+++ Scripts/test_imu.py
@@ -4,7 +4,7 @@ Make sure you are in the root directory of the project, not inside Scripts, and 
 For the pi, you will have to use python3
 """
 
-from Airbrakes.imu import IMU
+from airbrakes.imu import IMU
 
 # Should be checked before launch
 UPSIDE_DOWN = True
