+++ airbrakes/airbrakes.py
@@ -2,10 +2,10 @@
 
 from typing import TYPE_CHECKING
 
-from Airbrakes.imu import IMU, IMUDataPacket, RollingAverages
-from Airbrakes.logger import Logger
-from Airbrakes.servo import Servo
-from Airbrakes.state import StandByState, State
+from airbrakes.imu import IMU, IMUDataPacket, RollingAverages
+from airbrakes.logger import Logger
+from airbrakes.servo import Servo
+from airbrakes.state import StandByState, State
 
 if TYPE_CHECKING:
     import collections
@@ -60,7 +60,7 @@ class AirbrakesContext:
         data_packets: collections.deque[IMUDataPacket] = self.imu.get_imu_data_packets()
 
         # Logs the current state, extension, and IMU data
-        rolling_average = RollingAverages(data_packets.copy())
+        RollingAverages(data_packets.copy())
         # TODO: Compute state(s) for given IMU data
         self.logger.log(self.state.get_name(), self.current_extension, data_packets.copy())
